requirements.txt
torch>=1.9.0
numpy>=1.21.0
flask>=2.0.0
pyyaml>=5.4.0
qai-hub>=0.32.0

# Core ONNX requirements
onnx>=1.15.0
onnxruntime>=1.17.0

# For Snapdragon X Elite NPU support, you'll need:
# onnxruntime-qnn (Qualcomm specific package - install separately)
# 
# Install command for QNN runtime:
# pip install onnxruntime-qnn --index-url https://aiinfra.pkgs.visualstudio.com/PublicPackages/_packaging/onnxruntime-qnn/pypi/simple/
#
# Or download from Qualcomm developer portal:
# https://developer.qualcomm.com/software/qualcomm-ai-engine-direct

# Alternative GPU acceleration (if available)
# onnxruntime-gpu>=1.17.0